[MOB-7613] Move caching methods to separate file
